fix: #3067 updated ever and default seed
