--- apps/api/src/seed-default.ts
@@ -1,7 +1,0 @@
-import { seedDefault } from '@gauzy/core';
-import { pluginConfig } from './plugin-config';
-
-seedDefault(pluginConfig).catch((error: any) => {
-	console.log(error);
-	process.exit(1);
-});
