+++ apps/api/src/seed.ts
@@ -1,10 +1,7 @@
-import { seedEver } from '@gauzy/core';
+import { seedDefault } from '@gauzy/core';
 import { pluginConfig } from './plugin-config';
 
-/*
-* Ever seeder  
-*/
-seedEver(pluginConfig).catch((error: any) => {
+seedDefault(pluginConfig).catch((error: any) => {
 	console.log(error);
 	process.exit(1);
 });
