+++ packages/core/src/core/seeds/seed-jobs.ts
@@ -7,14 +7,11 @@ import { IPluginConfig } from '@gauzy/common';
 import { registerPluginConfig } from './../../bootstrap';
 import { SeedDataService } from './seed-data.service';
 import { SeederModule } from './seeder.module';
+
 /**
- * WARNING: Running this file will DELETE all data in your database
- * and generate and insert new, random data into your database.
- *
- * BE CAREFUL running this file in production env. It's possible to delete all production data.
- * SeedData checks if environment is in production or not by checking src/environments/environment.ts file configs.
- * If environment.production config is set to true, then the seeding process will only generate default roles and 2 default users.
- * */
+* WARNING: Running this file will generate and insert new, random jobs related data into your database.
+* 
+*/
 export async function seedJob(devConfig: Partial<IPluginConfig>) {
 	await registerPluginConfig(devConfig);
